Create the Card Object and logic for generating/randomizing cards
Made a comparable card object to represent each card in the deck.
This object is used in conjunction with service that allows a
standard deck to be created, as well as other functions such as
"dealing" and shuffling.

Start work on various scoring methods.  Currently working on
implementing the Royal Flush checks in such a way that the
code can be reused for straight flush etc.